data/validation/requirements.txt
build==1.2.1
cassandra-driver==3.29.1
docker==6.1.3
fonttools==4.50.0
fqdn==1.5.1
frozenlist==1.4.1
fs==2.4.16
fsspec==2024.3.1
geomet==0.2.1.post1
gitdb==4.0.11
greenlet==3.0.3
grpcio==1.62.2
grpcio-opentracing==1.1.4
grpcio-reflection==1.48.2
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.19.1
httpx==0.27.0
huggingface-hub==0.21.4
mistune==3.0.2
mlflow==2.11.3
numpy==1.26.4
protobuf==3.20.3
pyarrow==15.0.2
pycparser==2.21
pydantic==1.10.15
pydantic_core==2.18.1
Pygments==2.17.2
PyMySQL==1.0.3
pyparsing==2.4.7
pyproject_hooks==1.0.0
pyrsistent==0.20.0
pytest==8.1.1
torch==1.11.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
safetensors==0.4.2
schema==0.7.5
scikit-learn==1.4.1.post1
scipy==1.12.0
tokenizers==0.15.2
transformers==4.39.1
webencodings==0.5.1
websocket-client==1.7.0
yarl==1.9.4
zenml==0.55.5
zipp==3.18.1
